+++ src/ChunkColumn.ts
@@ -1,12 +1,13 @@
 
 /// <reference path="./global.d.ts" />
 import { Version, getChecksum } from "./format";
-import { BlockFactory } from './BlockFactory'
+import { BlockFactory, blockFactory } from './BlockFactory'
 import { Block } from "prismarine-block";
 import { StorageType, SubChunk } from './SubChunk'
 import nbt, { NBT } from "prismarine-nbt";
 import { Stream } from './Stream'
 import { BlobEntry, BlobStore, BlobType } from "./Blob";
+try { var v8 = require('v8') } catch {}
 
 const MIN_Y = 0
 const MAX_Y = 15
@@ -14,11 +15,11 @@ const MAX_Y = 15
 export class ChunkColumn {
   x: number; z: number
   version: number
-  sections: SubChunk[]
-  sectionsLen: number
+  sections: SubChunk[] = []
+  sectionsLen = 0
 
-  entities: NBT[]
-  tiles: { string?: NBT }
+  entities: NBT[] = []
+  tiles: { string?: NBT } = {}
 
   biomes?: Uint8Array
   heights?: Uint16Array
@@ -30,32 +31,38 @@ export class ChunkColumn {
   biomesUpdated = true
   biomesHash: Buffer | null
 
-  constructor(version: Version, x, z) {
-    this.version = version
+  factory = blockFactory
+
+  constructor(colVersion: Version, x: number, z: number) {
+    this.version = colVersion
     this.x = x
     this.z = z
-    this.sections = []
-    this.entities = []
-    this.tiles = {}
-    this.sectionsLen = 0
   }
 
-  getBlock(sx: int, sy: int, sz: int): Block {
-    let y = sy >> 4
-    let sec = this.sections[this.minY + y]
-    if (sec) return sec.getBlock(sx, sy & 0xf, sz)
-    return BlockFactory.getPBlockFromStateID(0)
+  getBlock({ x, y, z }): Block {
+    let Y = y >> 4
+    let sec = this.sections[this.minY + Y]
+    if (sec) return sec.getBlock(x, y & 0xf, z)
+    return this.factory.getPBlockFromStateID(0)
   }
 
-  setBlock(sx: int, sy: int, sz: int, block: Block) {
-    let y = sy >> 4
-    if (y < this.minY || y > this.maxY) return
-    let sec = this.sections[this.minY + y]
+  setBlock({ x, y, z }, block: Block) {
+    let Y = y >> 4
+    if (Y < this.minY || Y > this.maxY) return
+    let sec = this.sections[this.minY + Y]
     while (!sec) {
-      this.addSection(new SubChunk(this.version, this.sections.length))
-      sec = this.sections[this.minY + y]
+      this.addSection(new SubChunk(this.factory, this.version, this.sections.length))
+      sec = this.sections[this.minY + Y]
     }
-    return sec.setBlock(sx, sy & 0xf, sz, block)
+    return sec.setBlock(x, y & 0xf, z, block)
+  }
+
+  getBlockStateId(pos) {
+    return this.getBlock(pos)?.stateId
+  }
+
+  setBlockStateId(pos, runtimeId: number) {
+    return this.sections[0].setBlockID(pos.l || 0, pos.x, pos.y, pos.z, runtimeId)
   }
 
   addSection(section: SubChunk) {
@@ -72,7 +79,7 @@ export class ChunkColumn {
   }
 
   addBlockEntity(nbt) {
-    console.log('[wp] adding tile', nbt)
+    // console.log('[wp] adding tile', nbt)
     const x = nbt.value.x.value
     const z = nbt.value.z.value
     this.tiles[x + ',' + z] = nbt
@@ -90,11 +97,12 @@ export class ChunkColumn {
     return this.tiles
   }
 
-  getBiome(x, y, z) {
+  getBiome({ x, y, z }) {
     //todo
+    return 0
   }
 
-  setBiome(x, y, z, biome) {
+  setBiome({ x, y, z }, biome) {
     this.biomesUpdated = true
   }
 
@@ -179,7 +187,7 @@ export class ChunkColumn {
     // console.warn('Total Reading', sectionCount)
     for (let i = 0; i < sectionCount; i++) {
       // console.warn('Reading', i)
-      const section = new SubChunk(this.version, i, false)
+      const section = new SubChunk(this.factory, this.version, i, false)
       await section.decode(StorageType.Runtime, stream)
       this.sections.push(section)
     }
@@ -237,7 +245,7 @@ export class ChunkColumn {
       if (entry.type == BlobType.Biomes) {
         this.biomes = entry.buffer
       } else if (entry.type == BlobType.ChunkSection) {
-        const subchunk = new SubChunk(this.version, this.sectionsLen)
+        const subchunk = new SubChunk(this.factory, this.version, this.sectionsLen)
         await subchunk.decode(StorageType.NetworkPersistence, new Stream(entry.buffer))
         this.addSection(subchunk)
       } else {
@@ -246,6 +254,49 @@ export class ChunkColumn {
     }
 
     return misses
+  }
+
+
+  /* Serialization */
+  serialize() {
+    if (typeof v8 === 'undefined') {
+      throw Error('String serialization not yet supported')
+    } else {
+      const copy = { ...this, sections: [] }
+      delete copy.factory
+      for (const section of this.sections) {
+        const sec = { ...section }
+        delete sec.factory
+        copy.sections.push(v8.serialize(sec))
+      }
+      return v8.serialize(copy)
+    }
+  }
+
+  toJson() { return this.serialize() }
+
+  static deserialize(obj) {
+    if (typeof obj === 'string') {
+      // Oject.assign(this, JSON.parse(obj))
+      throw Error('String serialization not yet supported')
+    } else { // Buffer
+      const des = v8.deserialize(obj)
+      // @ts-ignore : we don't do anything special in the constructor, Object.assign should work
+      const chunk = new ChunkColumn()
+      Object.assign(chunk, des)
+      chunk.sections = []
+      for (const section of des.sections) {
+        // @ts-ignore : same for above
+        const s = new SubChunk(chunk.factory)
+        chunk.sections.push(Object.assign(s, v8.deserialize(section)))
+      }
+      // console.log('Des',obj,chunk)
+      return chunk
+    }
+  }
+
+  static fromJson(obj) {
+    return ChunkColumn.deserialize(obj)
   }
 }